refactor(design): move each button type to its own component
