--- libs/design/button/src/button-theme-variants/flat.scss
@@ -1,25 +1,0 @@
-@use 'sass:map';
-@use '../../../scss/theming';
-
-@mixin daff-flat-button-theme-variant(
-	$base-color,
-	$active-color
-) {
-	color: $base-color;
-
-	&:after {
-		background-color: $base-color;
-	}
-
-	&:hover {
-		color: theming.daff-text-contrast($base-color);
-	}
-
-	&:active {
-		color: theming.daff-text-contrast($active-color);
-
-		&:after {
-			background-color: $active-color;
-		}
-	}
-}
