--- libs/design/button/src/button-theme-variants/stroked.scss
@@ -1,28 +1,0 @@
-@use 'sass:map';
-@use '../../../scss/theming';
-
-@mixin daff-stroked-button-theme-variant(
-	$base-color,
-	$active-color
-) {
-	border: 1px solid $base-color;
-	color: $base-color;
-	background-color: transparent;
-
-	&:after {
-		background-color: $base-color;
-	}
-
-	&:hover {
-		color: theming.daff-text-contrast($base-color);
-	}
-
-	&:active {
-		border: 1px solid $active-color;
-		color: theming.daff-text-contrast($active-color);
-
-		&:after {
-			background-color: $active-color;
-		}
-	}
-}
