+++ libs/design/button/src/button.module.ts
@@ -3,16 +3,31 @@ import { NgModule } from '@angular/core';
 
 import { DaffPrefixSuffixModule } from '@daffodil/design';
 
-import { DaffButtonComponent } from './button/button.component';
+import { DaffButtonComponent } from './button/basic/button.component';
+import { DaffFlatButtonComponent } from './button/flat/flat.component';
+import { DaffIconButtonComponent } from './button/icon/icon.component';
+import { DaffRaisedButtonComponent } from './button/raised/raised.component';
+import { DaffStrokedButtonComponent } from './button/stroked/stroked.component';
+import { DaffUnderlineButtonComponent } from './button/underline/underline.component';
 
 /** @deprecated in favor of {@link DAFF_BUTTON_COMPONENTS} */
 @NgModule({
   imports: [
     CommonModule,
     DaffButtonComponent,
+    DaffFlatButtonComponent,
+    DaffIconButtonComponent,
+    DaffRaisedButtonComponent,
+    DaffStrokedButtonComponent,
+    DaffUnderlineButtonComponent,
   ],
   exports: [
     DaffButtonComponent,
+    DaffFlatButtonComponent,
+    DaffIconButtonComponent,
+    DaffRaisedButtonComponent,
+    DaffStrokedButtonComponent,
+    DaffUnderlineButtonComponent,
     DaffPrefixSuffixModule,
   ],
 })
