+++ libs/design/button/src/button/basic/button.component.scss
@@ -1,0 +1,29 @@
+@use '../button-base' as base;
+@use '../../../../scss/typography' as t;
+
+.daff-button {
+	@include base.daff-button-base();
+	@include base.daff-button-background(4px);
+	min-width: 96px;
+
+	&.daff-sm {
+		font-size: t.$small-font-size;
+		line-height: 2rem;
+		height: 2rem;
+		padding: 0 1rem;
+	}
+
+	&.daff-md {
+		font-size: t.$normal-font-size;
+		line-height: 3rem;
+		height: 3rem;
+		padding: 0 1.5rem;
+	}
+
+	&.daff-lg {
+		font-size: t.$medium-font-size;
+		line-height: 3.5rem;
+		height: 3.5rem;
+		padding: 0 1.5rem;
+	}
+}
