+++ libs/design/button/src/button/button-base.component.html
@@ -1,0 +1,12 @@
+<ng-container *ngIf="_prefix">
+  <ng-content select="[daffPrefix]"></ng-content>
+</ng-container>
+<ng-container *ngIf="!loading">
+  <span class="daff-button__content"><ng-content></ng-content></span>
+</ng-container>
+<ng-container *ngIf="loading">
+  <daff-loading-icon [diameter]="24"></daff-loading-icon>
+</ng-container>
+<ng-container *ngIf="_suffix">
+  <ng-content select="[daffSuffix]"></ng-content>
+</ng-container>
