--- libs/design/button/src/button/button.component.scss
@@ -1,207 +1,0 @@
-@use '../../../scss/typography' as t;
-@use '../../../scss/interactions';
-@use '../../../scss/layout';
-
-@mixin daff-button {
-	@include interactions.clickable;
-	display: inline-flex;
-	justify-content: center;
-	align-items: center;
-	gap: 8px;
-	appearance: none;
-	border: none;
-	border-radius: 4px;
-	position: relative;
-	text-align: center;
-	text-decoration: none;
-
-	&[disabled],
-	&.daff-button--disabled {
-		cursor: not-allowed;
-
-		&:hover {
-			&:after {
-				opacity: 0;
-			}
-		}
-	}
-
-	.daff-button__content,
-	.daff-prefix,
-	.daff-suffix {
-		z-index: 2;
-	}
-}
-
-@mixin daff-button-background($border-radius) {
-	&:after {
-		content: '';
-		border-radius: $border-radius;
-		position: absolute;
-		width: 100%;
-		height: 100%;
-		opacity: 0;
-		transition: opacity 300ms;
-	}
-
-	&:hover,
-	&:active {
-		&:after {
-			opacity: 1;
-		}
-	}
-}
-
-.daff-button,
-.daff-raised-button,
-.daff-stroked-button,
-.daff-flat-button {
-	@include daff-button();
-	min-width: 96px;
-
-	&.daff-sm {
-		font-size: t.$small-font-size;
-		line-height: 2rem;
-		height: 2rem;
-		padding: 0 1rem;
-	}
-
-	&.daff-md {
-		font-size: t.$normal-font-size;
-		line-height: 3rem;
-		height: 3rem;
-		padding: 0 1.5rem;
-	}
-
-	&.daff-lg {
-		font-size: t.$medium-font-size;
-		line-height: 3.5rem;
-		height: 3.5rem;
-		padding: 0 1.5rem;
-	}
-}
-
-.daff-button,
-.daff-flat-button {
-	@include daff-button-background(4px);
-}
-
-.daff-raised-button {
-	@include daff-button-background(0px);
-}
-
-.daff-stroked-button {
-	@include daff-button-background(3px);
-}
-
-.daff-flat-button {
-	background: none;
-	border: none;
-}
-
-.daff-icon-button {
-	@include daff-button();
-	background: none;
-	border: 0;
-	padding: 0;
-
-	&.daff-sm {
-		font-size: t.$small-font-size;
-		line-height: 2rem;
-		height: 2rem;
-		width: 2rem;
-	}
-
-	&.daff-md {
-		font-size: t.$normal-font-size;
-		line-height: 3rem;
-		height: 3rem;
-		width: 3rem;
-	}
-
-	&.daff-lg {
-		font-size: t.$medium-font-size;
-		line-height: 3.5rem;
-		height: 3.5rem;
-		width: 3.5rem;
-	}
-}
-
-.daff-underline-button {
-	@include daff-button();
-	background: transparent;
-	border: 0;
-	border-radius: 0;
-	line-height: 1.25rem;
-	overflow: hidden;
-	text-decoration: none;
-	vertical-align: middle;
-
-	&[disabled],
-	&.daff-button--disabled {
-		cursor: not-allowed;
-
-		&:hover,
-		&:active {
-			&::after {
-				animation: none;
-			}
-		}
-	}
-
-	&::after {
-		bottom: 0;
-		content: '';
-		height: 2px;
-		left: 0;
-		opacity: 1;
-		position: absolute;
-		width: 100%;
-	}
-
-	&:hover {
-		&::after {
-			animation: none;
-
-			@include layout.breakpoint(big-tablet) {
-				animation: underline-button-hover 700ms ease;
-			}
-		}
-	}
-
-	&.daff-sm {
-		font-size: t.$small-font-size;
-		height: 1.25rem;
-		padding: 0;
-	}
-
-	&.daff-md {
-		font-size: t.$normal-font-size;
-		height: 1.5rem;
-		padding: 0 0 4px;
-	}
-
-	&.daff-lg {
-		font-size: t.$medium-font-size;
-		height: 1.75rem;
-		padding: 0 0 8px;
-	}
-}
-
-@keyframes underline-button-hover {
-	0% {
-		transform: translateX(0);
-	}
-
-	50% {
-		transform: translateX(100%);
-	}
-
-	51% {
-		transform: translateX(-100%);
-	}
-
-	to {
-		transform: translateX(0);
-	}
-}
