+++ libs/design/button/src/button/flat/flat.component.spec.ts
@@ -1,0 +1,72 @@
+import {
+  Component,
+  DebugElement,
+} from '@angular/core';
+import {
+  waitForAsync,
+  ComponentFixture,
+  TestBed,
+} from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
+
+import { DaffFlatButtonComponent } from './flat.component';
+
+@Component({
+  template: `
+    <a daff-flat-button>Flat Link Button</a>
+    <button daff-flat-button>Flat Button</button>
+  `,
+  standalone: true,
+  imports: [
+    DaffFlatButtonComponent,
+  ],
+})
+
+class WrapperComponent {}
+
+describe('@daffodil/design/button | DaffFlatButtonComponent', () => {
+  let fixture: ComponentFixture<WrapperComponent>;
+  let wrapper: WrapperComponent;
+  let de: DebugElement;
+  let buttonDE: DebugElement;
+  let anchorDE: DebugElement;
+  let component: DaffFlatButtonComponent;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      imports: [
+        WrapperComponent,
+      ],
+    })
+      .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(WrapperComponent);
+    wrapper = fixture.componentInstance;
+    de = fixture.debugElement.query(By.css('[daff-flat-button]'));
+    component = de.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(wrapper).toBeTruthy();
+  });
+
+  describe('<daff-flat-button>', () => {
+    beforeEach(() => {
+      buttonDE = fixture.debugElement.query(By.css('button[daff-flat-button]'));
+      anchorDE = fixture.debugElement.query(By.css('a[daff-flat-button]'));
+    });
+
+    it('should add a class of `daff-flat-button` to the host element', () => {
+      expect(buttonDE.classes).toEqual(jasmine.objectContaining({
+        'daff-flat-button': true,
+      }));
+
+      expect(anchorDE.classes).toEqual(jasmine.objectContaining({
+        'daff-flat-button': true,
+      }));
+    });
+  });
+});
