+++ libs/design/button/src/button/icon/icon.component.spec.ts
@@ -1,0 +1,72 @@
+import {
+  Component,
+  DebugElement,
+} from '@angular/core';
+import {
+  waitForAsync,
+  ComponentFixture,
+  TestBed,
+} from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
+
+import { DaffIconButtonComponent } from './icon.component';
+
+@Component({
+  template: `
+    <a daff-icon-button>Icon Link Button</a>
+    <button daff-icon-button>Icon Button</button>
+  `,
+  standalone: true,
+  imports: [
+    DaffIconButtonComponent,
+  ],
+})
+
+class WrapperComponent {}
+
+describe('@daffodil/design/button | DaffIconButtonComponent', () => {
+  let fixture: ComponentFixture<WrapperComponent>;
+  let wrapper: WrapperComponent;
+  let de: DebugElement;
+  let buttonDE: DebugElement;
+  let anchorDE: DebugElement;
+  let component: DaffIconButtonComponent;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      imports: [
+        WrapperComponent,
+      ],
+    })
+      .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(WrapperComponent);
+    wrapper = fixture.componentInstance;
+    de = fixture.debugElement.query(By.css('[daff-icon-button]'));
+    component = de.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(wrapper).toBeTruthy();
+  });
+
+  describe('<daff-icon-button>', () => {
+    beforeEach(() => {
+      buttonDE = fixture.debugElement.query(By.css('button[daff-icon-button]'));
+      anchorDE = fixture.debugElement.query(By.css('a[daff-icon-button]'));
+    });
+
+    it('should add a class of `daff-icon-button` to the host element', () => {
+      expect(buttonDE.classes).toEqual(jasmine.objectContaining({
+        'daff-icon-button': true,
+      }));
+
+      expect(anchorDE.classes).toEqual(jasmine.objectContaining({
+        'daff-icon-button': true,
+      }));
+    });
+  });
+});
