+++ libs/design/button/src/button/raised/raised.component.spec.ts
@@ -1,0 +1,72 @@
+import {
+  Component,
+  DebugElement,
+} from '@angular/core';
+import {
+  waitForAsync,
+  ComponentFixture,
+  TestBed,
+} from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
+
+import { DaffRaisedButtonComponent } from './raised.component';
+
+@Component({
+  template: `
+    <a daff-raised-button>Raised Link Button</a>
+    <button daff-raised-button>Raised Button</button>
+  `,
+  standalone: true,
+  imports: [
+    DaffRaisedButtonComponent,
+  ],
+})
+
+class WrapperComponent {}
+
+describe('@daffodil/design/button | DaffRaisedButtonComponent', () => {
+  let fixture: ComponentFixture<WrapperComponent>;
+  let wrapper: WrapperComponent;
+  let de: DebugElement;
+  let buttonDE: DebugElement;
+  let anchorDE: DebugElement;
+  let component: DaffRaisedButtonComponent;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      imports: [
+        WrapperComponent,
+      ],
+    })
+      .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(WrapperComponent);
+    wrapper = fixture.componentInstance;
+    de = fixture.debugElement.query(By.css('[daff-raised-button]'));
+    component = de.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(wrapper).toBeTruthy();
+  });
+
+  describe('<daff-raised-button>', () => {
+    beforeEach(() => {
+      buttonDE = fixture.debugElement.query(By.css('button[daff-raised-button]'));
+      anchorDE = fixture.debugElement.query(By.css('a[daff-raised-button]'));
+    });
+
+    it('should add a class of `daff-raised-button` to the host element', () => {
+      expect(buttonDE.classes).toEqual(jasmine.objectContaining({
+        'daff-raised-button': true,
+      }));
+
+      expect(anchorDE.classes).toEqual(jasmine.objectContaining({
+        'daff-raised-button': true,
+      }));
+    });
+  });
+});
