+++ libs/design/button/src/button/stroked/stroked.component.spec.ts
@@ -1,0 +1,72 @@
+import {
+  Component,
+  DebugElement,
+} from '@angular/core';
+import {
+  waitForAsync,
+  ComponentFixture,
+  TestBed,
+} from '@angular/core/testing';
+import { By } from '@angular/platform-browser';
+
+import { DaffStrokedButtonComponent } from './stroked.component';
+
+@Component({
+  template: `
+    <a daff-stroked-button>Stroked Link Button</a>
+    <button daff-stroked-button>Stroked Button</button>
+  `,
+  standalone: true,
+  imports: [
+    DaffStrokedButtonComponent,
+  ],
+})
+
+class WrapperComponent {}
+
+describe('@daffodil/design/button | DaffStrokedButtonComponent', () => {
+  let fixture: ComponentFixture<WrapperComponent>;
+  let wrapper: WrapperComponent;
+  let de: DebugElement;
+  let buttonDE: DebugElement;
+  let anchorDE: DebugElement;
+  let component: DaffStrokedButtonComponent;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      imports: [
+        WrapperComponent,
+      ],
+    })
+      .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(WrapperComponent);
+    wrapper = fixture.componentInstance;
+    de = fixture.debugElement.query(By.css('[daff-stroked-button]'));
+    component = de.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(wrapper).toBeTruthy();
+  });
+
+  describe('<daff-stroked-button>', () => {
+    beforeEach(() => {
+      buttonDE = fixture.debugElement.query(By.css('button[daff-stroked-button]'));
+      anchorDE = fixture.debugElement.query(By.css('a[daff-stroked-button]'));
+    });
+
+    it('should add a class of `daff-stroked-button` to the host element', () => {
+      expect(buttonDE.classes).toEqual(jasmine.objectContaining({
+        'daff-stroked-button': true,
+      }));
+
+      expect(anchorDE.classes).toEqual(jasmine.objectContaining({
+        'daff-stroked-button': true,
+      }));
+    });
+  });
+});
