+++ libs/design/button/src/public_api.ts
@@ -1,5 +1,10 @@
 export { DaffButtonModule } from './button.module';
 export { DAFF_BUTTON_COMPONENTS } from './button';
 
-export { DaffButtonComponent } from './button/button.component';
+export { DaffButtonComponent } from './button/basic/button.component';
 export { DaffButtonSizableDirective } from './button/button-sizable.directive';
+export { DaffFlatButtonComponent } from './button/flat/flat.component';
+export { DaffRaisedButtonComponent } from './button/raised/raised.component';
+export { DaffStrokedButtonComponent } from './button/stroked/stroked.component';
+export { DaffUnderlineButtonComponent } from './button/underline/underline.component';
+export { DaffIconButtonComponent } from './button/icon/icon.component';
